packages/python/clerk-auth/pyproject.toml
[project]
name = "clerk-auth"
version = "0.1.0"
description = "Clerk JWT verification (JWKS + FastAPI deps) for Paperwork backends"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Paperwork Labs" }]
dependencies = [
    "fastapi>=0.110.0,<1.0",
    "httpx>=0.27.0",
    "pydantic>=2.7,<3",
    "python-jose[cryptography]>=3.3.0",
]

[project.optional-dependencies]
test = [
    "cryptography>=42",
    "mypy>=1.11",
    "pytest>=8",
    "pytest-asyncio>=0.21",
    "pytest-cov>=5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/clerk_auth"]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "DTZ", "SIM", "RUF", "ARG", "PIE", "C4"]

[tool.ruff.lint.per-file-ignores]
"src/clerk_auth/dependencies.py" = ["B008"]
"tests/**" = [
    "ARG001",
    "ARG002",
    "B008",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "--cov=clerk_auth --cov-report=term-missing --cov-fail-under=85"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["jose.*"]
ignore_missing_imports = true
